requirements.txt
keras
theano
flask
gevent
pillow
requests
